requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
torch>=1.12.0
torchvision>=0.13.0
